check Exam in update and create, remove return exam or notfound exception

+++ src/models/exam/exam.service.ts
@@ -7,6 +7,14 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { AppCustomLogger } from '../../app.custom.logger';
 import { Subject } from '../subject/entities/subject.entity';
 
+function isExamUndefined(createExamDto: CreateExamDto) {
+  return createExamDto.name === undefined || createExamDto.weighting === undefined;
+}
+
+function isExamEmpty(createExamDto: CreateExamDto) {
+  return createExamDto.name === "" || createExamDto.weighting === "";
+}
+
 @Injectable()
 export class ExamService {
   private readonly logger = new AppCustomLogger(ExamService.name);
@@ -14,8 +22,8 @@ export class ExamService {
   constructor(@InjectRepository(Exam) private readonly examRepository: Repository<Exam>, @InjectRepository(Subject) private readonly subjectRepository: Repository<Subject>){
   }
 
-  async create(createExamDto: CreateExamDto): Promise<Exam>{
-    if(createExamDto.name === undefined || createExamDto.weighting === undefined) {
+  async create(createExamDto: CreateExamDto): Promise<Exam | Error>{
+    if(isExamUndefined(createExamDto) || isExamEmpty(createExamDto)) {
       this.logger.warn(`Exam [create] does not have a name or weighting!`);
       throw new Error(`Exam [create] does not have a name or weighting!`);
     }
@@ -30,7 +38,7 @@ export class ExamService {
     return this.examRepository.find({ relations: ['modulpart'] });
   }
 
-  async findOne(id: number): Promise<Exam | undefined> {
+  async findOne(id: number): Promise<Exam | NotFoundException> {
     const entity = await this.examRepository.findOne({where: {id}, relations: ['modulpart']});
     if(!entity) {
       this.logger.warn(`Exam with id ${id} not found`);
@@ -39,11 +47,14 @@ export class ExamService {
     return entity;
   }
 
-  async update(id: number, updateExamDto: UpdateExamDto): Promise<Exam> {
+  async update(id: number, updateExamDto: UpdateExamDto): Promise<Exam | NotFoundException | Error> {
     const existingExam = await this.examRepository.findOne({where: {id}});
     if(!existingExam) {
       this.logger.warn(`Exam with id ${id} not found`);
       throw new NotFoundException(`Exam with id ${id} not found`);
+    }else if(isExamUndefined(updateExamDto) || isExamEmpty(updateExamDto)){
+        this.logger.warn(`Exam [update] does not have a name or weighting!`);
+        throw new Error(`Exam [update] does not have a name or weighting!`);
     }
     existingExam.name = updateExamDto.name;
     existingExam.weighting = updateExamDto.weighting;
@@ -51,13 +62,14 @@ export class ExamService {
     return this.examRepository.save(existingExam);
   }
 
-  async remove(id: number): Promise<void> {
-    const deleteResult = await this.examRepository.delete(id);
-    if(deleteResult.affected === 1) {
-      this.logger.log(`Exam with id ${id} deleted`);
-      //return {message: `Exam with id ${id} deleted`, status: 200};
-    }else{
+  async remove(id: number): Promise<Exam | NotFoundException> {
+    const findDeletedExam = await this.examRepository.findOne({where: {id}});
+    if(!findDeletedExam) {
+      this.logger.warn(`Exam with id ${id} not found`);
       throw new NotFoundException(`Exam with id ${id} not found`);
+    }else{
+      await this.examRepository.delete(id)
+      return findDeletedExam;
     }
   }
 }
